LayoutTests/imported/w3c/web-platform-tests/css/css-position/animations/right-composition-expected.txt


FAIL Compositing: property <right> underlying [50px] from add [100px] to add [200px] at (-0.3) should be [120px] assert_equals: expected "120px " but got "70px "
FAIL Compositing: property <right> underlying [50px] from add [100px] to add [200px] at (0) should be [150px] assert_equals: expected "150px " but got "100px "
FAIL Compositing: property <right> underlying [50px] from add [100px] to add [200px] at (0.5) should be [200px] assert_equals: expected "200px " but got "150px "
FAIL Compositing: property <right> underlying [50px] from add [100px] to add [200px] at (1) should be [250px] assert_equals: expected "250px " but got "200px "
FAIL Compositing: property <right> underlying [50px] from add [100px] to add [200px] at (1.5) should be [300px] assert_equals: expected "300px " but got "250px "
FAIL Compositing: property <right> underlying [100px] from add [10px] to add [2px] at (-0.5) should be [114px] assert_equals: expected "114px " but got "14px "
FAIL Compositing: property <right> underlying [100px] from add [10px] to add [2px] at (0) should be [110px] assert_equals: expected "110px " but got "10px "
FAIL Compositing: property <right> underlying [100px] from add [10px] to add [2px] at (0.5) should be [106px] assert_equals: expected "106px " but got "6px "
FAIL Compositing: property <right> underlying [100px] from add [10px] to add [2px] at (1) should be [102px] assert_equals: expected "102px " but got "2px "
FAIL Compositing: property <right> underlying [100px] from add [10px] to add [2px] at (1.5) should be [98px] assert_equals: expected "98px " but got "- 2px "
FAIL Compositing: property <right> underlying [10%] from add [100px] to add [20%] at (-0.3) should be [calc(130px + 4%)] assert_equals: expected "calc ( 4 % + 130px ) " but got "100px "
FAIL Compositing: property <right> underlying [10%] from add [100px] to add [20%] at (0) should be [calc(100px + 10%)] assert_equals: expected "calc ( 10 % + 100px ) " but got "100px "
FAIL Compositing: property <right> underlying [10%] from add [100px] to add [20%] at (0.5) should be [calc(50px + 20%)] assert_equals: expected "calc ( 20 % + 50px ) " but got "calc ( ( 100px * 0.5 ) + ( 20 % * 0.5 ) ) "
FAIL Compositing: property <right> underlying [10%] from add [100px] to add [20%] at (1) should be [30%] assert_equals: expected "30 % " but got "20 % "
FAIL Compositing: property <right> underlying [10%] from add [100px] to add [20%] at (1.5) should be [calc(-50px + 40%)] assert_equals: expected "calc ( 40 % - 50px ) " but got "20 % "
FAIL Compositing: property <right> underlying [50px] from add [100px] to replace [200px] at (-0.3) should be [135px] assert_equals: expected "135px " but got "70px "
FAIL Compositing: property <right> underlying [50px] from add [100px] to replace [200px] at (0) should be [150px] assert_equals: expected "150px " but got "100px "
FAIL Compositing: property <right> underlying [50px] from add [100px] to replace [200px] at (0.5) should be [175px] assert_equals: expected "175px " but got "150px "
PASS Compositing: property <right> underlying [50px] from add [100px] to replace [200px] at (1) should be [200px] 
FAIL Compositing: property <right> underlying [50px] from add [100px] to replace [200px] at (1.5) should be [225px] assert_equals: expected "225px " but got "250px "
FAIL Compositing: property <right> underlying [100px] from add [100px] to add [auto] at (-0.3) should be [200px] assert_equals: expected "200px " but got "100px "
FAIL Compositing: property <right> underlying [100px] from add [100px] to add [auto] at (0) should be [200px] assert_equals: expected "200px " but got "100px "
PASS Compositing: property <right> underlying [100px] from add [100px] to add [auto] at (0.5) should be [auto] 
PASS Compositing: property <right> underlying [100px] from add [100px] to add [auto] at (1) should be [auto] 
PASS Compositing: property <right> underlying [100px] from add [100px] to add [auto] at (1.5) should be [auto] 
